decouple api.ContentServer implementation package from bbolt dependency
For moby/buildkit#886
